refactor: move niche types to specific components
