--- types/offset-grid-element.ts
@@ -1,5 +1,0 @@
-export interface OffsetGridElement {
-  heading: string;
-  text: string;
-  icon: JSX.Element;
-}
