--- types/sitemap-element.ts
@@ -1,4 +1,0 @@
-export interface SitemapElement {
-  text: string;
-  href: string;
-}
